added tightness test in make_wf_interval
make_wf_interval calls make_untight_wf_interval, and then tightens
the result.  However, I believe that the original interval will
always be tight in the first place, since it's constructed by
min/max'ing two probability distributions--therefore every value
in the interval is in fact the value for some prob dist in the
interval, i.e. it is tight.  I want to check that so I'm continuing
to tighten the intervals but then testing whether they are actually
already tight.